authlab/core.py
# authlab/core.py

import os
import json
import time
import secrets
from datetime import datetime

from flask import (request, session, jsonify)

# --- .env autoload (dev convenience) ---
try:
    from dotenv import load_dotenv
    load_dotenv()
except Exception:
    pass

# --- Secrets from environment ---

ADMIN_PWHASH = os.getenv("ADMIN_PWHASH")
if not ADMIN_PWHASH:
    raise RuntimeError("ADMIN_PWHASH is not set (provide via environment or .env)")

SECRET_KEY = os.getenv("SECRET_KEY")
if not SECRET_KEY:
    raise RuntimeError("SECRET_KEY is not set (provide via environment or .env)")

# --- API error catalog ---

API_ERRORS = {
    "unauthorized": ("Login required", 401),
    "csrf_bad": ("Invalid CSRF token", 400),
    "bad_json": ("Expected application/json", 415),
    "empty": ("Message required", 400),
    "ratelimited": ("Too many requests", 429),
    "invalid_param": ("Bad parameter", 400),
    "invalid_range": ("Invalid range", 400),
    "invalid_sort_by": ("Invalid sort_by", 400),
    "invalid_sort_dir": ("Invalid sort_dir", 400),
    # + for global handlers:
    "not_found": ("Resource not found", 404),
    "method_not_allowed": ("Method not allowed", 405),
    "server_error": ("Internal server error", 500),
}

API_PRODUCTS_BUCKET = os.getenv("API_PRODUCTS_BUCKET", "api_products")
API_NOTES_BUCKET = os.getenv("API_NOTES_BUCKET", "api_notes")

# --- Rate-Limit config (fixed-window) ---

WINDOW_SEC  = int(os.getenv("WINDOW_SEC", 10))
MAX_ATTEMPTS = int(os.getenv("MAX_ATTEMPTS", 2))
RATE_BUCKET = os.getenv("RATE_BUCKET", "default")
RATE_STATE  = {}  # rate_key - {"start": int, "count": int}

# --- MFA config ---

ADMIN_MFA_ENABLED = (os.getenv("ADMIN_MFA_ENABLED", "false").lower() == "true")
ADMIN_MFA_SECRET  = os.getenv("ADMIN_MFA_SECRET")
MFA_WINDOW        = int(os.getenv("MFA_WINDOW", 1))
MFA_BUCKET        = os.getenv("MFA_BUCKET", "login_mfa")

if ADMIN_MFA_ENABLED and not ADMIN_MFA_SECRET:
    raise RuntimeError("ADMIN_MFA_ENABLED=true, but ADMIN_MFA_SECRET is missing")

# --- Users (store only hashes) ---

USERS = {
    "admin": {
        "password_hash": ADMIN_PWHASH,
        "mfa_enabled":  ADMIN_MFA_ENABLED,
        "mfa_secret":   ADMIN_MFA_SECRET,
    }
}

# --- Toggles ---

XSS_R_STATE = os.getenv("XSS_R_STATE", "safe").lower()   # reflected: poc|safe
XSS_S_STATE = os.getenv("XSS_S_STATE", "safe").lower()   # stored: poc|safe
SQLI_STATE  = os.getenv("SQLI_STATE",  "safe").lower()   # sqli: poc|safe
IDOR_STATE  = os.getenv("IDOR_STATE",  "safe").lower()   # idor: poc|safe

APP_ENV  = os.getenv("APP_ENV", "dev").lower()       # dev | prod
DEV_MODE = os.getenv("DEV_MODE", "false").lower() == "true"

if APP_ENV == "prod" and DEV_MODE:
    raise RuntimeError("DEV_MODE must be OFF in production")

# --- Guestbook state (in-memory) ---

GUESTBOOK   = []
MAX_MSG_LEN = int(os.getenv("MAX_MSG_LEN", 500))
NEXT_MSG_ID = 1

# --- Logs ---

LOG_DIR  = "logs"
LOG_FILE = os.path.join(LOG_DIR, "authlab.log")
os.makedirs(LOG_DIR, exist_ok=True)

def now_utc_iso():
    """Return current UTC time in ISO8601 with Z suffix."""
    return datetime.utcnow().isoformat() + "Z"

def client_ip():
    """Best-effort client IP from Flask request."""
    return request.remote_addr or "-"

def log_attempt(username, user_exists, result, reason, route=None, meta=None):
    """Append one structured authlab log record to authlab.log."""
    rec = {
        "ts": now_utc_iso(),
        "ip": client_ip(),
        "username": username,
        "user_exists": bool(user_exists),
        "result": result,
        "reason": reason,
        "route": route,
        "meta": meta,
    }
    with open(LOG_FILE, "a", encoding="utf-8") as f:
        f.write(json.dumps(rec, ensure_ascii=False) + "\n")

# --- Rate-limit helper (fixed window) ---

def rl_check_and_hit(rate_key, window_sec, max_attempts, now=None):
    """
    Fixed-window rate limit.

    Returns (allowed: bool, retry_after_seconds: int).
    If allowed=False - retry_after_seconds >= 1, attempt is NOT counted.
    If allowed=True  - attempt already counted (count++).
    """
    if now is None:
        now = int(time.time())

    state = RATE_STATE.get(rate_key)
    if state is None or now >= state["start"] + window_sec:
        state = {"start": now, "count": 0}
        RATE_STATE[rate_key] = state

    if state["count"] >= max_attempts:
        retry_after = (state["start"] + window_sec) - now
        if retry_after < 1:
            retry_after = 1
        return False, retry_after

    state["count"] += 1
    return True, 0


def json_ok(data, status=200, headers=None):
    """Uniform successful JSON response."""
    resp = jsonify(data)
    resp.status_code = status
    if headers:
        for k, v in headers.items():
            resp.headers[k] = v
    return resp


def json_err(code, message, status=400, details=None, headers=None):
    """Unified error JSON: { error: { code, message, details } }."""
    body = {"error": {"code": code, "message": message}}
    if details is not None:
        body["error"]["details"] = details
    resp = jsonify(body)
    resp.status_code = status
    if headers:
        for k, v in headers.items():
            resp.headers[k] = v
    return resp


def api_error(code, details=None):
    """Shortcut to build an error from API_ERRORS catalog."""
    msg, status = API_ERRORS[code]
    return json_err(code, msg, status=status, details=details)


def parse_int(val, default, min_v, max_v):
    """Safe int parser with bounds (for limit/offset)."""
    try:
        x = int(val)
    except Exception:
        return default
    return max(min_v, min(max_v, x))

def parse_float_or_none(val):
    """Parse float or return None on empty/invalid."""
    if val is None or val == "":
        return None
    try:
        return float(val)
    except (TypeError, ValueError):
        return None
    
def require_auth_json():
    """
    Ensure API user is authenticated.

    Returns (user, None) on success or (None, error_response).
    Supports DEV_MODE bypass with DEV_API_KEY for lab only.
    """
    user = session.get("user")
    if user:
        return user, None

    if DEV_MODE:
        dev_key = os.getenv("DEV_API_KEY")
        auth = request.headers.get("Authorization", "")
        if dev_key and auth.startswith("Bearer ") and auth[7:].strip() == dev_key:
            log_attempt("admin", True, "api_auth", "dev_api_key", route=request.path, meta=None)
            return "admin", None

    return None, api_error("unauthorized")

def ensure_csrf_token():
    """
    Ensure session['csrf_token'] exists (create if missing).

    Shared between HTML and API.
    """
    token = session.get("csrf_token")
    if not token:
        token = secrets.token_hex(32)
        session["csrf_token"] = token
    return token

def require_csrf_header():
    """
    For JSON POST from browser require X-CSRF-Token == session['csrf_token'].
    """
    expected = ensure_csrf_token()
    provided = request.headers.get("X-CSRF-Token")
    return provided and provided == expected
