requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
pyotp==2.9.0
schemathesis==4.6.4
